refactor: qc-core Coupon 도메인 패키지 정리
