+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/QuestionPaymentProcessor.java
@@ -1,9 +1,9 @@
 package com.eager.questioncloud.application.api.payment;
 
-import com.eager.questioncloud.core.domain.coupon.Coupon;
-import com.eager.questioncloud.core.domain.coupon.CouponRepository;
-import com.eager.questioncloud.core.domain.coupon.UserCoupon;
-import com.eager.questioncloud.core.domain.coupon.UserCouponRepository;
+import com.eager.questioncloud.core.domain.coupon.infrastructure.CouponRepository;
+import com.eager.questioncloud.core.domain.coupon.infrastructure.UserCouponRepository;
+import com.eager.questioncloud.core.domain.coupon.model.Coupon;
+import com.eager.questioncloud.core.domain.coupon.model.UserCoupon;
 import com.eager.questioncloud.core.domain.payment.QuestionPayment;
 import com.eager.questioncloud.core.domain.payment.QuestionPaymentOrderRepository;
 import com.eager.questioncloud.core.domain.payment.QuestionPaymentRepository;
