+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/coupon/infrastructure/CouponJpaRepository.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.coupon;
+package com.eager.questioncloud.core.domain.coupon.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
