+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/coupon/infrastructure/UserCouponEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.coupon;
+package com.eager.questioncloud.core.domain.coupon.infrastructure;
 
+import com.eager.questioncloud.core.domain.coupon.model.UserCoupon;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
