+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/coupon/infrastructure/UserCouponRepositoryImpl.java
@@ -1,8 +1,10 @@
-package com.eager.questioncloud.core.domain.coupon;
+package com.eager.questioncloud.core.domain.coupon.infrastructure;
 
-import static com.eager.questioncloud.core.domain.coupon.QCouponEntity.couponEntity;
-import static com.eager.questioncloud.core.domain.coupon.QUserCouponEntity.userCouponEntity;
+import static com.eager.questioncloud.core.domain.coupon.infrastructure.QCouponEntity.couponEntity;
+import static com.eager.questioncloud.core.domain.coupon.infrastructure.QUserCouponEntity.userCouponEntity;
 
+import com.eager.questioncloud.core.domain.coupon.dto.AvailableUserCoupon;
+import com.eager.questioncloud.core.domain.coupon.model.UserCoupon;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import com.querydsl.core.types.Projections;
